add func check_file

+++ file_manage.py
@@ -25,19 +25,31 @@ class File_Manager():
         return file
 
     @classmethod
+    def check_file(self, file_path: str, byte_count: int) -> bool:
+
+        if os.path.exists(file_path):
+            if os.path.getsize(file_path) == byte_count:
+                return False
+
+        return True
+
+    @classmethod
     def write_file(self, file_path: str, file_content: bytes) -> None:
 
         path = file_path[:-(len(file_path.split("\\")[-1]))]
         pathlib.Path(path).mkdir(parents=True, exist_ok=True)
 
-        file = self.open_file(file_path, write=True)
+        if self.check_file(file_path, len(file_content)):
+            file = self.open_file(file_path, write=True)
 
-        if file > 0:
-            os.truncate(file, 0)
+            if file > 0:
+                os.truncate(file, 0)
 
-        os.write(file, file_content)
-        print(f"EXTRACT: {file_path}")
-        os.close(file)
+            os.write(file, file_content)
+            print(f"EXTRACT: {file_path}")
+            os.close(file)
+        else:
+            pass
 
     @classmethod
     def open_memory(self, file: int, length: int, offset: int = 0) -> mmap:
